requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
Augmentor==0.2.9
cryptography==2.1.4
future==0.18.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
idna==2.6
joblib==1.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
Markdown==3.1.1
numpy==1.16.4
opencv-python==4.5.4.60
Pillow==8.3.2
protobuf==3.8.0
pycrypto==2.6.1
PyGObject==3.26.1
python-apt==1.6.4
pyxdg==0.25
scikit-learn==0.24.2
scipy==1.2.2
SecretStorage==2.3.1
six==1.11.0
sklearn==0.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
threadpoolctl==3.0.0
tqdm==4.62.3
Werkzeug==0.15.4
wrapt==1.11.2
